Refactor Splash Screen
