+++ lib/navigators/RootNavigation.js
@@ -1,6 +1,6 @@
 import { createStackNavigator, createBottomTabNavigator } from 'react-navigation'
 
-import SplashScreen from 'circles-mobile/lib/components/SplashScreen'
+import SplashContainer from 'circles-mobile/lib/containers/SplashContainer'
 import TermsConditionsScreen from 'circles-mobile/lib/components/Onboarding/OnboardingTerms'
 import AvatarScreen from 'circles-mobile/lib/components/Onboarding/OnboardingAvatar'
 import PhoneScreen from 'circles-mobile/lib/components/Onboarding/OnboardingPhone'
@@ -79,7 +79,7 @@ export const OrgHomeNavigator = createStackNavigator({
 
 export const IntroNavigator = createStackNavigator({
   Splash: {
-    screen: SplashScreen
+    screen: SplashContainer
   },
   Terms: {
     screen: TermsConditionsScreen
